update telegeram user id to Long

+++ modules/core/src/main/java/org/mycontroller/standalone/settings/TelegramBotSettings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2018 Jeeva Kandasamy (jkandasa@gmail.com)
+ * Copyright 2015-2022 Jeeva Kandasamy (jkandasa@gmail.com)
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -45,7 +45,7 @@ public class TelegramBotSettings {
 
     private String token;
 
-    private Integer id;
+    private Long id;
     private Boolean isBot;
     private String firstName;
     private String username;
